MILESTONE: flask app on 8081

+++ etl3_pdf2text.py
@@ -1,15 +1,18 @@
-import glob, fitz
+import glob, fitz, os
 from pprint import pprint
 
 #!mkdir TEXTOS
 for ix, fn in enumerate(list(glob.glob('PROGRAMAS/*.pdf'))):
     if ix%50==25:
         print(ix,fn)
-    fo = fitz.open(fn)
-    text = [page.getText() for page in fo]
-    fo.close()
-    tfn = fn.replace('PROGRAMAS','TEXTOS').replace('.pdf','.txt')
-    fw = open(tfn, 'w')
-    for page in text:
-        pprint(page, fw)
-    fw.close()
+    tfn = fn.replace('PROGRAMAS','TEXTOS/TODOS').replace('.pdf','.txt')
+    if os.path.exists(tfn):
+        pass
+    else:
+        fo = fitz.open(fn)
+        text = [page.getText() for page in fo]
+        fo.close()
+        fw = open(tfn, 'w')
+        for page in text:
+            pprint(page, fw)
+        fw.close()
